requirements.txt
# Core dependencies for AI Architect Bot
rich>=13.0.0              # Beautiful terminal formatting and progress bars
python-dotenv>=1.0.0      # Environment variable management
requests>=2.31.0          # HTTP client for API calls

# Optional dependencies for enhanced functionality
colorama>=0.4.6           # Cross-platform colored terminal text (Windows support)
